+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### Version 4.0.43
 
+* `api-scanner`: Compatibility fix for Gradle 5.x.
+
 ### Version 4.0.42
 
 * `cordformation`: Add Jolokia agent to the `cordaRuntime` configuration to prevent the `cordapp` plugin from including it inside the CorDapp.